third_party/blink/web_tests/platform/generic/fast/js/vardecl-preserve-parameters-expected.txt

This test verifies var declarations inside a function don't stomp function arguments.

PASS: x should be 1 and is.
